requirements.txt
PyMuPDF>=1.26.0
pdfplumber>=0.11.0
pytest>=8.0.0
Pillow>=10.0.0
